third_party/blink/web_tests/fast/dom/Document/createElement-invalid-names.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Test various valid name for elements according to the definition of DOM Level 2.")

var nameList = [
                // Invalid first Char
                '0ascii',
                '.Ascii',
                '-Ascii',
                ' ascii',
                '	ascii',
                // Invalid other char
                'asc i',
                'asc	i',
                ];

for (var i in nameList)
    shouldThrow("document.createElement('" + nameList[i] +"')");
</script>
</body>
</html>
